Add predicate using HMAsset attributes
